Add more tests to getFullRect()
This commit fleshes out tests of the remaining functionality for
getFullRect(). While working on these, I discovered a bug in how we
handle the margins of the topmost nodes. Instead of treating them as
extensions of specific boxes, we are simply adding them onto the boxes,
so they are having an exaggerated effect. I don't have time to fix this
now, so I decided to write a failing test as part of this commit.