Add ability to use publicDirectories
This is useful if you for instance have custom fonts. We've had this as
`public_directories` in the ruby version of happo.

I'll try to come up with a test case for this as well, likely porting
over what we had from happo_spec.rb. but that will be for another
commit.