Extract uploading logic into new class
I want to add some test coverage to `happojs upload`. To make it easier
to mock out aws and the s3 service, I'm moving all upload-related code
into a new class. A nice side-effect of this is that it will get easier
to add another upload-service. All we have to do is implement the same
interface and provide a way to switch between different services in
configuration.

No actual tests were added as part of this commit. I didn't have time.
I'll come back here later.