Error out if example renders nothing
I found this while trying out the beta release on the brigade codebase.
We have a component that doesn't render anything visible below a certain
screensize. In the past, we've silently just saved a 1x1 pixel image (I
think), but I think it's better to yell about it. Happo shouldn't be the
tool to validate that components are hidden.