Wait for animation frame before taking screenshot
On initial testing of the beta release (happo 3.0.0-beta.3) I saw a lot
of spurious diffs coming from css gradients. It was as if Firefox
rendered linear-gradient in two steps. If we ended up taking the
screenshot in between these two steps, there was a slight difference
between the gradients. By wrapping rendering in a call to
`requestAnimationFrame` it looks like the problem goes away.

Thanks to @jede for suggesting this when I expressed frustration over
the spuriousness in an informal chat.