Fix travis build
I forgot that we were running `npm run build` from travis. In 0362bed I
renamed `build` as `prepublish`, which broke travis builds. The good
news is that `prepublish` is run automatically as part of a raw `npm
install`, which is also the default installation script for travis. So
we can just remove our custom installation procedure.