Compare snapshots in the background
By not waiting for a snapshot to be compared to a previous one while
processing examples, we can make runVisualDiffs a little faster. My
non-scientific tests of this on the brigade codebase show a 5-7%
reduction in total `happo run` time.

I still believe the most impactful thing we can do for performance at
this point is to only screenshot the visible content (instead of taking
a full-screen screenshot and then cropping). But that will come later.