Flatten tests
Reduced indirection improves the readability of these tests. Normally
this type of refactoring reduces the DRYness of the tests as well, but
that is not the case here. Even if it were, it would be worth it to
maximize the locality of the tests to make them readable.

I think we could actually go further with this, but I'll leave that for
another time.