Limit screenshot to example box
For a long while now, we've been screenshotting the whole page, then
cropping it to only the relevant box on the backend. This is bad for
performance, and we believe that the most impactful thing we can do for
happo performance right now is to make sure that screenshots are
pre-cropped.

I started looking at the webdriver spec, and noticed that it has a
section about screenshotting an element [1]. Selenium doesn't mention
this in docs, but you can use this method.

  driver.findElement(By.id('foobar')).then((element) => {
    element.takeScreenshot().then((screenshot) => {
      // do something
    });
  });

When testing this out, it didn't work in firefox (using the firefox
driver). It didn't work in Chrome either. Nor in Safari. It turns out
that the only two webdrivers supporting this method is Edge (since a
while back) and geckodriver (recently).

Performance is definitely improved. On the brigade codebase, a full run
used to take roughly 2 min 45 seconds. After this commit, we're down to
less than 2 minutes! The larger the viewport, the bigger the win. We run
most our examples on 640x888 and 320x444 viewports, where it looks like
we're not really gaining much. But the larger viewports have a
noticeable perf improvement.

[1]: https://www.w3.org/TR/webdriver/#take-element-screenshot