Wait for background images as well as regular images
With some versions of Firefox, there's a race condition between loading
background images and generating a screenshot. Sometimes the background
image is available, sometimes it's not. This leads to spurious diffs. I
started seeing this more after trying out geckodriver, but it looks like
it's present in the current firefoxdriver as well, as visualized by this
diff:
https://react-primitives-happo.s3-us-west-1.amazonaws.com/487da291eaa1c64410200dbf4fc91462/index.html

There's no DOM event to listen to for background image loads, so I'm
implementing a technique that involves getComputedStyle for each node in
an example. We worried that this would be too slow, but preliminary
tests show that this only adds about 15ms per snapshot (tested on the
Brigade codebase with 1000+ components), which I think is
okay. If we at some point in the future would want to optimize this, we
could add a configuration option to allow people to turn off the
behavior.

This geckodriver issue discusses this problem in more depth:
https://github.com/mozilla/geckodriver/issues/367