Avoid erroring on 0x0 nodes
While doing this, we discovered some issues with getFullRect() that
prevented it from working properly if there are no nodes or if the nodes
are 0x0.

The testing here was pretty tricky to figure out. I kept running into
failures due to async timeouts being hit. Error messaging was poor, and
the only reasonable way to troubleshoot was by dropping console.logs
everywhere. It makes me wonder if we are perhaps not quite using our
Promises correctly.

Even more annoying, once I fixed our test for this case, it seems to
cause the next few tests to break.