Fix flaky tests
Our runVisualDiffs test has been very flaky for a while. Timeouts keep
happening, and it's been really hard to track down exactly what's going
on. I played around a little with memoizing a driver in between runs,
and found that tests were really fast and a lot more stable.

I had to fix naming of the short/tall examples so that they shared the
same name, and after that all specs were passing.