Handle errors in getImageFromStream
I was trying out happo locally and once saw the following:

      throw er; // Unhandled 'error' event
      ^

Error: There are some read requests waitng on finished stream
  at ChunkStream._end (/path/to/node_modules/pngjs/lib/chunkstream.js:100:7)
  at ChunkStream.end (/path/to/node_modules/pngjs/lib/chunkstream.js:86:10)
  at PNG.end (/path/to/node_modules/pngjs/lib/png.js:97:16)
  at ReadStream.onend (_stream_readable.js:517:10)
  at ReadStream.g (events.js:286:16)
  at emitNone (events.js:91:20)
  at ReadStream.emit (events.js:185:7)
  at endReadableNT (_stream_readable.js:934:12)
  at _combinedTickCallback (internal/process/next_tick.js:74:11)
  at process._tickDomainCallback (internal/process/next_tick.js:122:9)

This doesn't give me a lot to go on, but I'm wondering if we can get
some more information by handling stream errors and rejecting the
promise, the next time this happens.

I added the error handling in two places because it needs to happen
before and after the pipe in order to provide the clearest messaging.
See http://stackoverflow.com/a/22389498/18986