Don't fail `happo upload` when there are no results
This brings back old behavior from the ruby gem. When `happo upload`
detects no changes (nothing to upload), we will now exit with a zero
status code. No URL will be logged. This is what the old tool used to
do, but I tried to be smart about changing it.