Move S3 configuration into AWS.S3 initialization
These should be equivalent, but I am running into permission problem
when trying to upload to S3 in one project. The only thing that makes
sense to me is that our configuration must not be getting wired up
correctly. I've seen both styles of configuration, but since this one
seems simpler and makes more sense to me, I'm going to give it a go.