Assume bucket exists when getting 403 while checking for existence
I am running into an issue where I don't have access to check on the
existence of a bucket or create a new bucket, so the upload process
bails with this error. This started happening once we rewrote in
JavaScript from Ruby.

One of the differences might be we moved from Amazon's v2 of the SDK to
v3. We found an issue on the PHP SDK that looks related

  https://github.com/aws/aws-sdk-php/issues/1029

Our hypothesis is that the bucket actually exists but we don't have
access to determine this even though we have access to write to the
bucket. If this is the case, we are going to try proceeding blindly and
hoping that the upload works.