Add support for multiple background-images
As pointed out by @lencioni in a previous commit, we need to handle the
case where background-image is a list of urls.

https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Background_and_Borders/Using_CSS_multiple_backgrounds

I've extracted the url-extracting into a new function and added tests to
verify its behavior.

While here, I also fixed an issue with assuming NodeList has a forEach
method, using the same technique as we do a few lines up in the file.