Don't exlude node-modules from being transformed for Jest
We were seeing this error when running tests in Node <= 5:

  SyntaxError: Unexpected token ...

This error was coming from trying to parse this file
node_modules/selenium-webdriver/lib/http.js

  function toExecuteAtomCommand(command, atom, ...params) {

According to the readme, node 4 and 5 are supported, which makes me a
little confused about having to transform them using babel.

https://www.npmjs.com/package/selenium-webdriver#node-support-policy

Tests seem to run a little slower after this change, but only on the
first run.