Mention geckodriver in README
Even though we bundle geckodriver, it will only work for Mac OS and
Linux x64. I don't know how big of an issue this is, so I'm not worrying
too much about documenting the fact that you might have to install
geckodriver separately. I'll wait until that becomes an actual problem.
In the meantime, we can at least mention it and link to the repo.