Make `closeDriver` always resolve, never reject
I was seeing this error when running happo in CentOS:

  UnhandledPromiseRejectionWarning: Unhandled promise rejection (rejection
  id: 1): Error: Command failed: ps l

This was coming from src/server/closeDriver.js, which has the sole purpose
of working around a bug in marionette [1]. By making it log the error
while resolving the promise we can avoid the unhandled rejection. I'm
not 100% sure why the `ps l` command fails, but looking at the processes
on that machine I see no orphaned firefoxes running, so it's most likely
safe to ignore.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1310992