Upgrade selenium-webdriver and geckodriver
I noticed geckodriver just released 0.15.0, which I was hoping would fix
the issues we've been seeing with closing the driver. It requires the
latest selenium-webdriver, but it turns out that it's imcompatible with
it after all. Sigh.

https://github.com/SeleniumHQ/selenium/issues/3625

In the end, I bumped us up to geckodriver 0.14.0 at least, which seems
to work well with the latest selenium.

I also had to bump us to firefox v52 to get selenium 3.3.0 working.