Remove closeDriver.js workaround
Due to a bug in how geckodriver was closing, orphaned firefox instances
were left running in the background. We've had a workaround in place to
`kill` the orphaned process. I was never happy with how that ended up,
and even though I'm not sure whether the upstream bug has been fixed,
I'm removing the workaround. I can no longer reproduce this locally at
least - instead I'm getting ugly logs from the latest version of
ps-node when it kills the process.