Fix `happo debug firefox`
On running this command, we would provide you with a link to a "debug"
page, where we listed all the happo examples defined in your code. But
because of things having been moved around, it led to a page showing an
error (something like "`css` is undefined").

By referencing the files directly from happo-viewer we can fix this. And
with that change I was able to remove some duplication as well.