Load official geckodriver
When we first started using geckodriver, the released versions didn't
have support for screenshooting individual DOM elements. So we built our
own binaries and bundled them with the lib. At this point, geckodriver
has caught up and we no longer need the custom build.

The node-geckodriver [1] npm module simplifies downloading the right
geckodriver binary and putting it on the PATH to make it available for
Selenium.

[1] https://github.com/vladikoff/node-geckodriver