Remove polymorphism in waitForImagesToLoad
I added this to allow different code paths to call it with either a
string or an image element. I did this because it made it possible to
grab the crossorigin attribute from an image (which we need when
preloading). It ended up being messy so I cleaned up the function and
updated call sites. As an added bonus, if an <img> with a srcset has a
crossorigin attribute, it is now honored (before, only <img src> would
pass it along).